+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - Introduced new packages which serve as a wrapper for different Mediator libraries.
 - **ZEA.Communications.Messaging.Abstractions**
     - Added the possibility to add metadata to a publish command.
+- **ZEA.Applications.Logging.Metadata.Abstractions**
+- **ZEA.Applications.Logging.Metadata.MVC**
 
 ### Deprecated